feat(mongo): add index on @id field
